% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_threshold.R
\name{above_threshold}
\alias{above_threshold}
\title{Compute amount above a threshold}
\usage{
above_threshold(x, threshold)
}
\arguments{
\item{x}{A dataframe containing time-series data.}

\item{threshold}{A numeric value specifying the threshold.}
}
\value{
A dataframe indicating the amount exceeding the giving threshold.
}
\description{
Compute amount above a threshold
}
