% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_dist_2shp.R
\name{ave_dist_2shp}
\alias{ave_dist_2shp}
\title{WRITTEN BY LUIS BECERRA ON 28/06/2024. Updated on 14/03/2025
This function calculates the average distance (in m) of each \code{ID} element in \code{target} to the object \code{shape_input}
It uses distanceFromPoints which works better, faster. st_distance function yields weird results
Works better for UTM now}
\usage{
ave_dist_2shp(country_ext, target, shape_input, UTM, res_raster, name_dist)
}
\arguments{
\item{country_ext}{defines the whole extension of the country, the country borders. It could use GADM as extension}

\item{target}{is the shapefile (adm-div such as communes, villages or HH coordinates) for which the
average distance will be calculated.}

\item{shape_input}{is the shapefile for which a raster of distances will be calculated (eg. roads, rivers, etc)

\code{target}, \code{shape_input} and \code{country_ext} have to be in "sf-data.frame" in WGS84}

\item{UTM}{corresponds to the country's location in the UTM coordinate system.
As for calculating distances it is more convenient to use meters instead of degrees,
and the coordinate system needs to be transformed into the UTM of the country, which
is more precise around the equator.
Otherwise, it uses the default EPSG:3857 (Web Mercator),
"+proj=merc +datum=WGS84 +units=m" which holds at the global level}

\item{res_raster}{is the resolution in meters for the raster of distances that will be created.
A higher resolution increase the time of processing}

\item{name_dist}{name of the average distance variable constructed. It will be
created as ave_dist_NAME with name_dist="NAME"}
}
\value{
A \link[base]{data.frame}, with the average distance variable calculated for each \code{ID}
}
\description{
WRITTEN BY LUIS BECERRA ON 28/06/2024. Updated on 14/03/2025
This function calculates the average distance (in m) of each \code{ID} element in \code{target} to the object \code{shape_input}
It uses distanceFromPoints which works better, faster. st_distance function yields weird results
Works better for UTM now
}
\examples{
ave_dist_rivers <- ave_dist_2shp(country_ext=gadm3,target=polygons_gns,shape_input=rivers_fao, UTM=31,res_raster=10000,name_dist="rivers")

}
\seealso{
\code{\link[raster:distanceFromPoints]{raster::distanceFromPoints()}}
}
