% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_point_density_2shp.R
\name{ave_point_density_2shp}
\alias{ave_point_density_2shp}
\title{WRITTEN BY LUIS BECERRA ON 28/06/2024. Updated on 14/03/2025
This function calculates the average density of points for each \code{ID} element of \code{target}to the object \code{shape_input}}
\usage{
ave_point_density_2shp(
  country_ext,
  target,
  shape_input,
  res_raster,
  name_density
)
}
\arguments{
\item{country_ext}{defines the whole extension of the country, the country borders. It could use GADM as extension}

\item{target}{is the shapefile (adm-div such as communes, villages or HH coordinates) for which the
average point density will be calculated.

\code{target}, \code{shape_input} and \code{country_ext} have to be in "sf-data.frame" in WGS84}

\item{shape_input}{is the shapefile from which to calculate points density (eg. roads, rivers, etc)}

\item{res_raster}{is the resolution in degrees for the raster of distances that will be created and applied
to `shape_input' (0.1 could be equivalent to 10km or 0.1*100000=10000m )
A higher resolution increase the time of the process}

\item{name_density}{name of the average density variable constructed. It will be
created as ave_den_NAME with name_density="NAME"}
}
\value{
A \link[base]{data.frame}, with the average density variable calculated for each \code{ID}
}
\description{
WRITTEN BY LUIS BECERRA ON 28/06/2024. Updated on 14/03/2025
This function calculates the average density of points for each \code{ID} element of \code{target}to the object \code{shape_input}
}
