% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/below_threshold.R
\name{below_threshold}
\alias{below_threshold}
\title{Compute amount below a threshold}
\usage{
below_threshold(x, threshold)
}
\arguments{
\item{x}{A dataframe containing time-series data.}

\item{threshold}{A numeric value specifying the threshold.}
}
\value{
A dataframe indicating the amount below the threshold.
}
\description{
Compute amount below a threshold
}
