% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_to_clock.R
\name{character_to_clock}
\alias{character_to_clock}
\title{Convert a character vector to a clock object}
\usage{
character_to_clock(x)
}
\arguments{
\item{x}{A character vector representing dates in various formats, with
underscores or periods.}
}
\value{
A vector of parsed date-time objects of class \code{clock} from the
\code{clock} package.
}
\description{
This function takes a character vector, replaces underscores or periods with
dashes, removes any leading 'X', and parses the resulting string as a
date-time object.
}
\examples{
# Example input with underscores
x1 <- c("X2022_01_01", "X2023_02_02")
character_to_clock(x1)

# Example input with periods
x2 <- c("X2022.01.01", "X2023.02.02")
character_to_clock(x2)
}
