% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic4economist.R
\docType{_PACKAGE}
\name{climataic4economist}
\alias{climataic4economist}
\title{climataic4economist Package}
\description{
Contains functions for extracting weather data based on survey location and
to construct weather extreme indicators
}
\author{
Jan Martin Rossi
}
