% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic4economist.R
\docType{package}
\name{climatic4economist}
\alias{climatic4economist-package}
\alias{climatic4economist}
\title{climatic4economist Package}
\description{
Contains functions for extracting weather data based on survey location and
to construct weather extreme indicators
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmr-fao/climatic4economist}
  \item Report bugs at \url{https://github.com/jmr-fao/climatic4economist/issues}
}

}
\author{
Jan Martin Rossi
}
