% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spell.R
\name{compute_spell}
\alias{compute_spell}
\title{Compute spell length for consecutive days}
\usage{
compute_spell(condition, threshold = 2)
}
\arguments{
\item{condition}{A boolean vector indicating threshold exceedance.}

\item{threshold}{A numeric value specifying the minimum spell length (default
is 2).}
}
\value{
A numeric vector indicating the length of spells where conditions
hold.
}
\description{
Compute spell length for consecutive days
}
