% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_water_balance.R
\name{compute_water_balance}
\alias{compute_water_balance}
\title{Compute Water Balance}
\usage{
compute_water_balance(pre, pet)
}
\arguments{
\item{pre}{A data frame containing precipitation values with ID, x_cell, and y_cell columns.}

\item{pet}{A data frame containing potential evapotranspiration values with the same structure as \code{pre}.}
}
\value{
A data frame with the same structure as \code{pre}, where the values represent the water balance.
}
\description{
This function calculates the water balance by summing precipitation (\code{pre}) and
potential evapotranspiration (\code{pet}) for each corresponding cell.
}
\examples{
pre <- data.frame(ID = 1:3, x_cell = c(10, 20, 30), y_cell = c(40, 50, 60), `2020` = c(100, 120, 140))
pet <- data.frame(ID = 1:3, x_cell = c(10, 20, 30), y_cell = c(40, 50, 60), `2020` = c(-50, -60, -70))
compute_water_balance(pre, pet)
}
