% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_std_index.R
\name{extr_std_index}
\alias{extr_std_index}
\title{Extract Standardized Index Based on Thresholds}
\usage{
extr_std_index(df, l_thrshld = NULL, u_thrshld = NULL)
}
\arguments{
\item{df}{A data frame containing the data to be standardized.}

\item{l_thrshld}{A list of lower threshold values. If NULL, lower threshold classification is skipped.}

\item{u_thrshld}{A list of upper threshold values. If NULL, upper threshold classification is skipped.}
}
\value{
A data frame containing the original data with additional standardized index columns.
}
\description{
This function computes standardized indices based on upper and lower thresholds.
If a value exceeds or falls below the respective threshold, it is classified accordingly.
}
\examples{
df <- data.frame(ID = 1:3, value = c(5, 10, 15))
l_thrshld <- list(8)
u_thrshld <- list(12)
extr_std_index(df, l_thrshld, u_thrshld)

}
