% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extr_abs_day.R
\name{find_extr_abs_day}
\alias{find_extr_abs_day}
\title{Identify absolute extreme weather events}
\usage{
find_extr_abs_day(df, u_thresh = NULL, l_thresh = NULL, unit = NULL)
}
\arguments{
\item{df}{A dataframe containing time-series data with column names
representing dates and a variable \code{ID} that identifies the unique units.}

\item{u_thresh}{A numeric vector specifying upper threshold values. Optional,
default is NULL.}

\item{l_thresh}{A numeric value specifying the lower threshold. Optional,
default is NULL.}

\item{unit}{A character string specifying the unit of measure of the weather
observations, (default: "unit"). E.g. "mm" for precipitation, "C" for
temperature.}
}
\value{
A dataframe containing:
\itemize{
\item for each threshold a logical column indicating if the day was more extreme
events than the threshold.
\item for each threshold a numeric column indicating the amount of excess with
respect to the threshold.
}
}
\description{
This function detects extreme weather events based on user-defined
absolute upper and lower threshold values. It identifies in which days the
weather observation is more extreme than the thresholds and
calculates the amount of the excess.
}
\details{
For each \code{ID} unit the function computes indicators for values exceeding the
upper and lower thresholds.

The argument \code{unit} does not have implication for the computations but it
labels the result.

The user can specify either or both the \code{u_thresh} and the \code{l_thresh}.
If none are specified the result contains just the weather observations.
}
\examples{
df <- data.frame(ID = 1:3,
                 X2022.12 = c(10, 15, 5),
                 X2022.06 = c(12, 18, 3))
find_extr_abs_day(df, u_thresh = c(10, 20), l_thresh = 0.1)
}
