% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_folder.R
\name{find_folder}
\alias{find_folder}
\title{Find a Subfolder Within a Root Directory}
\usage{
find_folder(folder_name, root = "Users")
}
\arguments{
\item{folder_name}{A character string giving the name of the subfolder to
search for.}

\item{root}{A character string specifying the name of the root directory
from which the search should begin. Defaults to \code{"Users"}.}
}
\value{
A character string containing the full path to the first matching sub folder.
The function returns \code{NULL} (or the value returned by
\code{climatic4economist::find_subfolder()}) if no matching folder is found.
}
\description{
Searches for a subfolder with a given name starting from a specified root
directory. The function first locates the root directory using
\code{climatic4economist::find_root_directory()} and then recursively searches
for the target subfolder using
\code{climatic4economist::find_subfolder()}.
}
\examples{
\dontrun{
find_folder("Documents")

find_folder(
  folder_name = "sp_repository",
  root = "Users"
)
}
}
