% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_root_directory.R
\name{find_root_directory}
\alias{find_root_directory}
\title{Find a Root Directory by Walking Upward}
\usage{
find_root_directory(start = getwd(), root_directories, ignore_case = TRUE)
}
\arguments{
\item{start}{Character. Starting path (default = current working directory).}

\item{root_directories}{Character vector. Names of folders to match (case-insensitive by default).}

\item{ignore_case}{Logical. If TRUE, the search is case-insensitive (default = TRUE).}
}
\value{
Character path to the found root directory, or \code{NULL} if not found.
}
\description{
Starting from a given path (default: current working directory), this function
walks up the directory tree until it finds a folder matching one of the specified names.
}
\examples{
\dontrun{
find_root_directory(start = getwd(), root_directories = c("OneDirive", ".git"))
}

}
