% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_above.R
\name{is_above}
\alias{is_above}
\title{Identify values above a threshold}
\usage{
is_above(x, threshold)
}
\arguments{
\item{x}{A dataframe containing time-series data.}

\item{threshold}{A numeric value specifying the threshold.}
}
\value{
A dataframe with a boolean column indicating whether each value
exceeds the threshold.
}
\description{
Identify values above a threshold
}
