% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_below.R
\name{is_below}
\alias{is_below}
\title{Identify values below a threshold}
\usage{
is_below(x, threshold)
}
\arguments{
\item{x}{A dataframe containing time-series data.}

\item{threshold}{A numeric value specifying the threshold.}
}
\value{
A dataframe with a boolean column indicating whether each value is
below the threshold.
}
\description{
Identify values below a threshold
}
