% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date.R
\name{is_date}
\alias{is_date}
\title{Check if a value is a valid date object}
\usage{
is_date(mydate)
}
\arguments{
\item{mydate}{character. A character string representing a date.}
}
\value{
A logical value indicating whether the input is a valid date.
}
\description{
This function checks whether a given value can be parsed as a valid date.
}
\examples{
is_date("2023-01-01")  # TRUE
is_date("not a date")  # FALSE
}
