% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_na_check.R
\name{mean_na_check}
\alias{mean_na_check}
\title{Compute the mean of a numeric vector while handling missing values}
\usage{
mean_na_check(x, replace = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{replace}{A numeric value to return if all elements in \code{x} are \code{NA}.
Defaults to \code{0}.}
}
\value{
A numeric value representing the mean of \code{x}, ignoring \code{NA} values.
If all values in \code{x} are \code{NA}, it returns \code{replace}.
}
\description{
This function calculates the mean of a numeric vector while handling \code{NA}
values. If all values in the vector are \code{NA}, it returns a specified
replacement value.
}
\examples{
# Example with no missing values
mean_na_check(c(1, 2, 3, 4, 5))  # Returns 3

# Example with missing values
mean_na_check(c(1, 2, NA, 4, 5))  # Returns 3

# Example where all values are NA
mean_na_check(c(NA, NA, NA))  # Returns 0

# Example with a custom replacement value
mean_na_check(c(NA, NA, NA), replace = -1)  # Returns -1
}
