% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_coord.R
\name{prepare_coord}
\alias{prepare_coord}
\title{Prepare ID for unique occurrences}
\usage{
prepare_coord(df, lat_var, lon_var)
}
\arguments{
\item{df}{A data frame}

\item{lat_var}{\if{html}{\out{<data-masking>}} variable with the latitude coordinate}

\item{lon_var}{\if{html}{\out{<data-masking>}} variable with the longitude coordinate}
}
\value{
The same data frame \code{df} with an additional \code{ID} variable that
identifies the unique locations.
}
\description{
This this function adds an \code{ID} columns that identifies unique occurrence
based on unique coordinates.
}
\examples{
prepare_coord(survey, gps_lat, gps_lon)
}
