% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_by_dates.R
\name{select_by_dates}
\alias{select_by_dates}
\title{Select time observations based on date range}
\usage{
select_by_dates(df, from = NULL, to = NULL)
}
\arguments{
\item{df}{A data frame with time-series data in wide format.}

\item{from}{character. A string representing the start date (optional).}

\item{to}{character. A string representing the end date (optional).}
}
\value{
A data frame filtered based on the specified date range.
}
\description{
This function selects columns based on a specified date range.
}
\details{
The function requires both or just one between the starting date,
\code{from}, and the end date \code{to}. If both are provide the the function select
between the two dates. If only \code{from} is provided the function selects all
date after. If only \code{to} is provided the function selects all date
before. If none are provided returns the data frame unchanged.
}
\examples{
df <- data.frame(X2022.12 = c(100, 200),
                 X2022.06 = c(150, 250))

select_by_dates(df, from = "2022-06-01", to = "2022-12-31")
}
