% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spi.R
\name{spi_wide}
\alias{spi_wide}
\title{Compute the SPI}
\usage{
spi_wide(df_list, scale = 1)
}
\arguments{
\item{df_list}{A data frame. It must contains an \code{value} variable and a
variable \code{name} with the date of observations.}

\item{scale}{an integer, representing the time scale at which the SPEI / SPI
will be computed.}
}
\value{
A \link[tibble]{tbl_df}, with the SPI values in wide format, i.e.
each column is an observation in time.
}
\description{
Compute the Standardized Precipitation Index (SPI) in wide format.
}
\examples{
spi_wide(data, scale = 1)
}
\seealso{
\link[SPEI:spi]{SPEI::spi}
}
