% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_l.R
\name{substitute_l}
\alias{substitute_l}
\title{Replace Values Below a Threshold with NA}
\usage{
substitute_l(x, threshold)
}
\arguments{
\item{x}{A numeric vector.}

\item{threshold}{A numeric value specifying the lower bound. Values below
this threshold will be replaced with \code{NA_real_}.}
}
\value{
A numeric vector with values below the threshold replaced by
\code{NA_real_}.
}
\description{
This function replaces values in a numeric vector that are below a given
threshold with \code{NA_real_}.
}
\examples{
vec <- c(2, 5, 8, 1, 10)
substitute_l(vec, threshold = 5)
# Returns: c(NA, 5, 8, NA, 10)
}
