% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_u.R
\name{substitute_u}
\alias{substitute_u}
\title{Replace Values Above a Threshold with NA}
\usage{
substitute_u(x, threshold)
}
\arguments{
\item{x}{A numeric vector.}

\item{threshold}{A numeric value specifying the upper bound. Values above this threshold will be replaced with \code{NA_real_}.}
}
\value{
A numeric vector with values above the threshold replaced by \code{NA_real_}.
}
\description{
This function replaces values in a numeric vector that are above a given threshold with \code{NA_real_}.
}
\examples{
vec <- c(2, 5, 8, 1, 10)
substitute_u(vec, threshold = 5)
# Returns: c(2, 5, NA, 1, NA)
}
