% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_date.R
\name{to_date}
\alias{to_date}
\title{Convert a String to a Date-Like Format}
\usage{
to_date(x)
}
\arguments{
\item{x}{A character vector containing strings to be converted.}
}
\value{
A character vector with transformed strings.
}
\description{
This function transforms a string by replacing underscores (\verb{_}) and dots (\code{.})
with hyphens (\code{-}) and removing a leading "X" if present. It is useful for
formatting column names or other identifiers into a more readable date format.
}
\examples{
to_date("X2023_05_01")
"2023-05-01"

to_date("X2023.12.31")
"2023-12-31"

to_date("2023/12/31")
"2023-12-31"
}
