% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_stata_format.R
\name{to_stata_format}
\alias{to_stata_format}
\title{Convert Variable Names to Stata-Compatible Format}
\usage{
to_stata_format(x)
}
\arguments{
\item{x}{A character vector of variable names or dates.}
}
\value{
A character vector with modified names.
}
\description{
This function reformat a string (e.g., a date or column name) to be
compatible with Stata variable naming conventions. It substitute dashes with
underscore, remove dots and ensures that names starting with a number are
prefixed with "X" (since Stata does not allow variable names to start with
numbers).
}
\examples{
to_stata_format("2024-01-01") # Returns "X2024_01_01"
to_stata_format("var.1")      # Returns "var1"
to_stata_format("123_test")   # Returns "X123_test"
}
