% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_par.R
\name{calc_par}
\alias{calc_par}
\title{Calculate Aggregated Parameters}
\usage{
calc_par(df, pars, prefix = NULL)
}
\arguments{
\item{df}{A data frame containing at least an "ID" column and numeric columns
with years as names.}

\item{agg_fns}{A function or a named list or a named vector of functions to
be applied for aggregation.}
}
\value{
A data frame with aggregated statistics for each ID.
}
\description{
This function computes summary statistics for each unique ID by applying
a set of aggregation functions to the selected numeric columns.
}
\examples{
df <- data.frame(
  ID = c(1, 1, 2, 2, 3),
  `2001` = c(10, 15, 20, 25, 30),
  `2002` = c(5, 10, 15, 20, 25)
)
calc_par(df, list(mean = mean, sum = sum))

}
