% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pct_day.R
\name{calc_pct_day}
\alias{calc_pct_day}
\title{Compute daily percentile for specific location and month}
\usage{
calc_pct_day(
  df,
  iteracation = NULL,
  p,
  l_thresh = NULL,
  u_thresh = NULL,
  yearly = FALSE
)
}
\arguments{
\item{df}{A dataframe containing an ID column and date-based columns.}

\item{iteracation}{optional character to be print before computation. Usually,
it is the name of the object on which the function is applied. This is useful
when the function is used inside an apply family function to keep track of the
iterations.}

\item{p}{A numeric vector specifying the percentiles to compute (values
between 0 and 1).}

\item{l_thresh}{(Optional) A numeric value specifying the lower threshold.
Values below this will be excluded.}

\item{u_thresh}{(Optional) A numeric value specifying the upper threshold.
Values above this will be excluded.}

\item{yearly}{(optional) Logical, if TRUE calculate also yearly daily values.}
}
\value{
A dataframe with computed percentile-based day values for each month,
with column names prefixed by "day_".
}
\description{
This function calculates the specified percentiles of daily values for each
month, optionally filtering values based on lower and upper
thresholds. Optionally it calculates the specified percentiles of daily values
for the entire year.
}
\examples{
calc_pct_day(df, p = c(0.25, 0.5, 0.75), l_thresh = 60, u_thresh = 0.1)
}
