% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spei.R
\name{compute_spei}
\alias{compute_spei}
\title{Compute the SPEI}
\usage{
compute_spei(df, time_scale, iteracation = NULL)
}
\arguments{
\item{df}{A data frame. It must contains an \code{ID} variable and the
precipitation observation in wide format, i.e. each column is a different
observation in time.}

\item{time_scale}{an integer, representing the time scale at which the SPEI
will be computed.}

\item{iteracation}{optional character to be print before computation. Usually,
it is the name of the object on which the function is applied. This is useful
when the function is used inside an apply family function to keep track of the
iterations.}
}
\value{
A \link[tibble]{tbl_df}, It must contains an \code{ID} variable and the
SPEI observation in wide format, i.e. each column is a different
observation in time and each row is a unique location.
}
\description{
Compute the Standardized Precipitation-Evapotranspiration Index (SPEI) for
unique locations. Under the hood the function performs parallel computation
to speed up the computation with \link[furrr:future_map]{furrr::future_map}.
}
\examples{
compute_spei(coord, time_scale = 1)
}
\seealso{
\link[SPEI:Drought-indices]{SPEI::spei}
}
