% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_with_buffer.R
\name{crop_with_buffer}
\alias{crop_with_buffer}
\title{Crop a Raster with a Buffer around a Vector Layer}
\usage{
crop_with_buffer(raster, vector, buffer = 0, buffer_unit = "native", ...)
}
\arguments{
\item{raster}{A \code{SpatRaster} object to be cropped.}

\item{vector}{A \code{SpatVector} object used to define the cropping extent.}

\item{buffer}{A numeric value specifying the buffer (in map units) to expand the cropping extent.
Default is 0 (no buffer).}

\item{...}{Additional arguments passed to \code{\link[terra:crop]{terra::crop()}}.}
}
\value{
A cropped \code{SpatRaster} object based on the buffered extent of the input vector layer.
}
\description{
Crops a \code{SpatRaster} object using the extent of a \code{SpatVector} object,
optionally adding a buffer around the vector extent before cropping.
}
\details{
The function computes the bounding box of the input vector and expands it in all directions
by the buffer amount. The raster is then cropped to this expanded extent.
}
\examples{
\dontrun{
library(terra)
r <- rast(system.file("ex/elev.tif", package = "terra"))
v <- vect(system.file("ex/lux.shp", package = "terra"))
cropped <- crop_with_buffer(r, v, buffer = 500)
plot(cropped)
}
}
