% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_day_index.R
\name{extr_day_index}
\alias{extr_day_index}
\title{Compute Extreme Day Indices for a Given Time Interval}
\usage{
extr_day_index(
  df,
  iteracation = NULL,
  interview,
  id,
  interval,
  n_lags = 0,
  extra_col = NULL
)
}
\arguments{
\item{df}{A data frame containing daily extreme values.}

\item{iteracation}{optional character to be print before computation. Usually,
it is the name of the object on which the function is applied. This is useful
when the function is used inside an apply family function to keep track of the
iterations.}

\item{interview}{A column name specifying the interview date for reference.}

\item{id}{A column name specifying the unique identifier for each unit}

\item{interval}{A string specifying the time interval (e.g., \code{"30days"},
\code{"3 months"}, \code{"1 year"}).}

\item{n_lags}{Integer. The maximum lag (in units of \code{interval}) to consider.
Default is 0.}
}
\value{
A data frame with:
\itemize{
\item \code{id}: The unique identifier.
\item \code{lag}: The time lag relative to the interview date.
\item Summarized extreme day counts (\verb{day_*}) and unit-based extreme values
(\verb{*_abv}, \verb{*_blw}).
}
}
\description{
This function calculates indices of extreme events occurring before an
interview date, aggregating them over a specified time interval.
}
\examples{
df <- data.frame(
  ID = c(1, 1, 1, 2, 2),
  date = as.character(c("2024-01-01", "2024-01-02", "2024-01-03", "2024-01-01", "2024-01-02")),
  day_abv_90p = c(0, 1, 1, 0, 0),
  temp_abv_90p = c(1, 2, 3, 1, 0)
)
extr_day_index(df, interview = "2024-02-01", id = ID, interval = "30days", n_lags = 1)
}
