% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_by_coord.R
\name{extract_by_coord}
\alias{extract_by_coord}
\title{Extract values from a SpatRaster}
\usage{
extract_by_coord(raster, coord, iteration = NULL, ...)
}
\arguments{
\item{raster}{A data frame. It must contains two variables named \code{lat} and
\code{lon}.}

\item{coord}{character. The coordinate reference system in one of the
following formats: WKT/WKT2, \if{html}{\out{<authority>}}:\if{html}{\out{<code>}}, or PROJ-string notation.
See \link[terra]{crs}.}

\item{iteration}{optional character to be print before computation. Usually,
it is the name of the object on which the function is applied. This is useful
when the function is used inside an apply family function to keep track of the
iterations.}

\item{...}{additional arguments to pass to \link[terra:extract]{terra::extract}.}
}
\value{
A  \link[tibble]{tbl_df}, which the extracted values for each
location and the coordinates of the locations. If the raster has multiple
layers, the values of the layer will be split by columns.
}
\description{
Extract values from a \link[terra]{SpatRaster} for a set of locations.
The locations must be a point \link[terra]{SpatVector}.
}
\examples{
georef_coord(survey, crs = "epsg:4326")
}
\seealso{
\code{\link[terra]{extract}}
}
