% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_by_poly.R
\name{extract_by_poly}
\alias{extract_by_poly}
\title{Extract and Aggregate Raster Values by Polygons}
\usage{
extract_by_poly(raster, poly, fn_agg = mean, categorical = FALSE, na_rm = TRUE)
}
\arguments{
\item{raster}{A \code{SpatRaster} object from the \code{terra} package.}

\item{poly}{A \code{SpatVector} object representing polygon features.}

\item{fn_agg}{A function name specifying how to aggregate raster values within each polygon.
Default is \code{mean}.}
}
\value{
A \code{tibble} with aggregated raster values, one row per polygon.
}
\description{
This function extracts raster values within polygon features and aggregates them
using a specified summary function (e.g., "mean", "sum", "median"). It returns
the extracted values as a tibble.
}
\examples{
r <- terra::rast(nrows=10, ncols=10)
terra::values(r) <- runif(ncell(r))
p <- vterra::ect(matrix(c(0,0,1,0,1,1,0,1,0,0), ncol=2, byrow=TRUE), type="polygons")

extract_by_poly(r, p)
}
\seealso{
\code{\link[terra]{extract}}
}
