% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cell_by_poly.R
\name{extract_cell_by_poly}
\alias{extract_cell_by_poly}
\title{Extract Raster Values by Polygon}
\usage{
extract_cell_by_poly(raster, poly, ...)
}
\arguments{
\item{raster}{A \code{SpatRaster} or \code{RasterLayer} object representing the raster
data to extract.}

\item{poly}{A \code{sf} or \code{SpatialPolygonsDataFrame} object representing the
polygon(s) for extraction.}

\item{...}{additional arguments to pass to \link[exactextractr:exact_extract]{exactextractr::exact_extract}.}
}
\value{
A data frame containing extracted values, along with coordinates
(\code{x}, \code{y}) and coverage fraction.
}
\description{
This function extracts raster values within a given polygon using the
\code{exactextractr} package.
}
\examples{
# Example raster
r <- terra::rast(nrows=100, ncols=100, xmin=0, xmax=10, ymin=0, ymax=10)
terra::values(r) <- runif(ncell(r), min = 0, max = 100)

# Example polygon
poly <- sf::st_as_sf(sf::st_buffer(st_sfc(sf::st_point(c(5, 5))), dist = 1))

# Extract raster values within the polygon
extract_by_poly(r, poly)
}
\seealso{
\code{\link[exactextractr]{exact_extract}}
}
