% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extr_rel_day.R
\name{find_extr_rel_day}
\alias{find_extr_rel_day}
\title{Identify relative extreme days}
\usage{
find_extr_rel_day(
  df,
  iteration = NULL,
  u_thresh = NULL,
  l_thresh = NULL,
  unit = "unit",
  id = NULL
)
}
\arguments{
\item{df}{A dataframe containing a unit identifier column and date-based
columns.}

\item{iteration}{optional character to be print before computation. Usually,
it is the name of the object on which the function is applied. This is useful
when the function is used inside an apply family function to keep track of the
iterations.}

\item{u_thresh}{(Optional) A dataframe containing upper percentile thresholds
for each ID and month. Typically computed using \code{calc_pct_day()}.}

\item{l_thresh}{(Optional) A dataframe containing lower percentile thresholds
for each ID and month. Typically computed using \code{calc_pct_day()}.}

\item{unit}{(Optional) A character string specifying the unit name to replace
"unit" in output columns.}

\item{id}{Optional character string naming the column that identifies the
units. When \code{NULL}, \code{ID} is used, or \code{ID_adm_div} when that is the only
one present.}
}
\value{
A dataframe with columns indicating whether values exceeded or fell
below percentile thresholds, along with the corresponding differences.
\itemize{
\item \code{day_abv_Xp}: Boolean indicating if weather exceeds the threshold
\code{Xp}.
\item \code{unit_abv_Xp}: The amount by which weather exceeds the upper threshold
\code{Xp}.
\item \code{day_abv_max_Xp}: Boolean indicating if weather exceeds the max
upper threshold among all the months.
\item \code{day_blw_Xp}: Boolean indicating if weather exceeds the lower threshold
\code{Xp}.
\item \code{unit_blw_Xp}: The amount by which weather exceeds the lower threshold
\code{Xp}.
\item \code{day_blw_max_Xp}: Boolean indicating if weather exceeds the min
lower threshold among all the months.
}
}
\description{
This function detects relative extreme values in a dataset based on
percentile thresholds. It identifies days where values exceed or fall below
specified percentiles and calculates the excess or deficit relative to these
thresholds.
}
