% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extr_spell.R
\name{find_extr_spell_rel}
\alias{find_extr_spell_rel}
\title{Identify relative extreme spells}
\usage{
find_extr_spell_rel(df, threshold, min_spell = 2)
}
\arguments{
\item{df}{A data frame containing precipitation data with an \code{ID} column and
daily precipitation values .}

\item{threshold}{A data frame containing dry spell percentile thresholds for
each \code{ID} and \code{month}, typically computed using \code{calc_pct_spell()}.}

\item{dry_treshold}{A numeric value (default = 0.1) defining the
precipitation threshold below which a day is considered dry.}
}
\value{
A data frame containing:
\itemize{
\item \code{ID}: Unique identifier for locations or individuals.
\item \code{date}: The date of the observation.
\item \code{dry_spell_abv_Xp}: Difference between observed dry spell duration and
threshold \code{Xp}, where \code{Xp} represents different percentiles.
}
}
\description{
This function detects extreme spells by comparing observed spell
duration with predefined percentile-based thresholds.
}
