% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lag.R
\name{find_lag}
\alias{find_lag}
\title{Compute the number of time intervals between two dates}
\usage{
find_lag(start, end, width = 1, unit = "year")
}
\arguments{
\item{start}{A \code{Date} or \code{POSIXt} object representing the start date.}

\item{end}{A \code{Date} or \code{POSIXt} object representing the end date.}

\item{width}{An integer specifying the width of the time interval. Defaults
to 1.}

\item{unit}{A character string specifying the unit of time to measure
(\code{"year"}, \code{"month"}, etc.). Defaults to \code{"year"}. Must be a valid unit
recognized by \link[lubridate:time_length]{lubridate::time_length}.}
}
\value{
An integer representing the number of complete intervals of the
specified width between the start and end dates.
}
\description{
This function calculates the number of complete time intervals (e.g., years
or months) between a start and an end date, based on a specified interval
width.
}
\examples{
# Example: Find the number of full years between two dates
find_lag(as.Date("2020-01-01"), as.Date("2025-01-01"))

# Example: Find the number of 6-month intervals
find_lag(as.Date("2020-01-01"), as.Date("2023-01-01"),
         width = 6, unit = "months")

}
\seealso{
\code{\link[lubridate:time_length]{lubridate::time_length()}} \code{\link[lubridate:interval]{lubridate::interval()}}
}
