% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef_coord.R
\name{georef_coord}
\alias{georef_coord}
\title{Georeference based on coordinates and coordinate reference system}
\usage{
georef_coord(df, geom, crs)
}
\arguments{
\item{df}{A data frame.}

\item{geom}{character. The field name(s) with the geometry data. Two names
for x and y coordinates of points.}

\item{crs}{character. The coordinate reference system in one of the following
formats: WKT/WKT2, \if{html}{\out{<authority>}}:\if{html}{\out{<code>}}, or PROJ-string notation. See
\link[terra]{crs}.}
}
\value{
A \link[terra]{SpatVector} of points with one variable \code{ID} as
location unique identifier.
}
\description{
This this function transform a data frame into a point
\link[terra]{SpatVector}.
}
\examples{
georef_coord(survey, crs = "epsg:4326")
}
\seealso{
\link[terra:vect]{terra::vect}
}
