% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_with_survey.R
\name{merge_with_survey}
\alias{merge_with_survey}
\title{Merge survey observations with other data.}
\usage{
merge_with_survey(survey, new_value, iteracation = NULL)
}
\arguments{
\item{survey}{A data frame. It must contains an \code{ID} for the merging.}

\item{new_value}{A data frame. The data to be merge with the
survey. It must contains an \code{ID} variable for the merging.}

\item{iteracation}{optional character to be print before computation. Usually,
it is the name of the object on which the function is applied. This is useful
when the function is used inside an apply family function to keep track of the
iterations.}
}
\value{
A \link[tibble]{tbl_df}, with variables from the survey and new
weather values.
}
\description{
The functions uses the variable \code{ID} as the key matching variable.
}
\examples{
spi_survey <- merge_with_survey(survey, spi_values)
}
\seealso{
\code{\link[dplyr:mutate-joins]{dplyr::full_join()}}
}
