% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_df.R
\name{quantile_df}
\alias{quantile_df}
\title{Compute percentiles and return as data frame}
\usage{
quantile_df(x, p, replace)
}
\arguments{
\item{x}{A numeric vector for which percentiles are computed.}

\item{p}{A numeric vector specifying the probabilities at which percentiles
should be computed (values between 0 and 1).}
}
\value{
A tibble (data frame) with the computed percentiles, where column
names are derived from the probability values (e.g., \verb{25p}, \verb{50p}, \verb{75p}).
}
\description{
This function calculates the specified percentiles of a numeric vector and
returns the result as a tidy dataframe.
}
\seealso{
\link{quantile_na_check}
}
