% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_na_check.R
\name{quantile_na_check}
\alias{quantile_na_check}
\title{Compute sample percentiles of a numeric vector while handling missing
values}
\usage{
quantile_na_check(x, p, replace = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{numeric vector of probabilities with values in [0, 1].}

\item{replace}{A numeric value to return if all elements in \code{x} are \code{NA}.
Defaults to \code{0}.}
}
\value{
A numeric value representing the standard deviation of \code{x}, ignoring
\code{NA} values. If all values in \code{x} are \code{NA}, it returns \code{replace}.
}
\description{
This function calculates sample percentiles corresponding to the given
probabilities. If all values in the vector are \code{NA}, it returns a
specified replacement value.
}
\seealso{
\code{\link[=quantile]{quantile()}}
}
