% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spei.R
\name{spei_wide}
\alias{spei_wide}
\title{Compute the SPEI}
\usage{
spei_wide(df_list, scale = 1)
}
\arguments{
\item{df_list}{A data frame. It must contains an \code{value} variable and a
variable \code{name} with the date of observations.}

\item{scale}{an integer, representing the time scale at which the SPEI
will be computed.}
}
\value{
A \link[tibble]{tbl_df}, with the SPEI values in wide format, i.e.
each column is an observation in time.
}
\description{
Compute the Standardized Precipitation-Evapotranspiration Index (SPEI) in
wide format.
}
\examples{
spei_wide(data, scale = 1)

}
\seealso{
\link[SPEI:Drought-indices]{SPEI::spei}
}
